+++ Tests/xcprojTests/PBXTargetDependencySpec.swift
@@ -7,11 +7,13 @@ final class PBXTargetDependencySpec: XCTestCase {
     var subject: PBXTargetDependency!
 
     override func setUp() {
-        subject = PBXTargetDependency(target: "target",
+        subject = PBXTargetDependency(name: "name",
+                                      target: "target",
                                       targetProxy: "target_proxy")
     }
 
     func test_init_initializesTheTargetDependencyWithTheCorrectAttributes() {
+        XCTAssertEqual(subject.name, "name")
         XCTAssertEqual(subject.target, "target")
         XCTAssertEqual(subject.targetProxy, "target_proxy")
     }
@@ -21,13 +23,14 @@ final class PBXTargetDependencySpec: XCTestCase {
     }
 
     func test_equals_shouldReturnTheRightValue() {
-        let one = PBXTargetDependency(target: "target", targetProxy: "target_proxy")
-        let another = PBXTargetDependency(target: "target", targetProxy: "target_proxy")
+        let one = PBXTargetDependency(name: "name", target: "target", targetProxy: "target_proxy")
+        let another = PBXTargetDependency(name: "name", target: "target", targetProxy: "target_proxy")
         XCTAssertEqual(one, another)
     }
 
     private func testDictionary() -> [String: Any] {
         return [
+            "name": "name",
             "target": "target",
             "targetProxy": "targetProxy",
             "reference": "reference"